+++ modin/tests/pandas/native_df_mode/test_indexing.py
@@ -19,6 +19,7 @@ import pytest
 
 import modin.pandas as pd
 from modin.config import NativeDataframeMode, NPartitions
+from modin.tests.pandas.native_df_mode.utils import eval_general_interop
 from modin.tests.pandas.utils import (
     RAND_HIGH,
     RAND_LOW,
@@ -27,7 +28,6 @@ from modin.tests.pandas.utils import (
     default_to_pandas_ignore_string,
     df_equals,
     eval_general,
-    eval_general_interop,
     test_data,
     test_data_keys,
     test_data_values,
